Fixes variable referenced before assignment

+++ src/eda_server/ruleset.py
@@ -115,8 +115,8 @@ async def activate_rulesets(
 
     elif deployment_type == "docker" or deployment_type == "podman":
 
+        container = None
         docker = aiodocker.Docker()
-
         host = "eda-server"
 
         try:
@@ -153,7 +153,8 @@ async def activate_rulesets(
             await container.start()
         except aiodocker.exceptions.DockerError as e:
             logger.error("Failed to start container: %s", e)
-            await container.delete()
+            if container is not None:
+                await container.delete()
             await docker.close()
             raise
